api/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
